OSIS-2481 Move pdf methods in a new file.
