+++ base/templates/education_group/pdf_content.html
@@ -1,4 +1,4 @@
-{% load education_group %}
+{% load education_group education_group_pdf %}
 {% load staticfiles %}
 {% load i18n %}
 <!doctype html>
