+++ base/tests/templatetags/test_build_pdf_tree.py
@@ -28,7 +28,7 @@ from django.test import TestCase
 
 from base.models.enums.learning_component_year_type import LECTURING, PRACTICAL_EXERCISES
 from base.models.enums.learning_unit_year_periodicity import BIENNIAL_ODD, BIENNIAL_EVEN
-from base.templatetags.education_group import pdf_tree_list
+from base.templatetags.education_group_pdf import pdf_tree_list
 from base.tests.factories.education_group_year import EducationGroupYearFactory
 from base.tests.factories.group_element_year import GroupElementYearFactory
 from base.tests.factories.learning_component_year import LearningComponentYearFactory
